+++ yes-read-api/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Add 'own' filter implementation to read API
 
 # 1.2.0 - 2025-10-17
 * Bump yousty-eventsourcing dependency to 15.0.1-alpha8